requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.4.0
chardet==4.0.0
charset-normalizer==2.0.12
colorama==0.4.4
discord-py-interactions @ git+https://github.com/interactions-py/library.git@00ebd1b896f7cc17415cebd04cfc1877a6badbc8
enhanced==3.2.1
frozenlist==1.3.0
idna==3.3
multidict==6.0.2
orjson==3.6.3
Pillow==9.0.1
pysos==1.2.6
python-dotenv==0.19.2
typing-extensions==4.0.1
yarl==1.7.2
